Convert config.sh to UNIX newlines before sourcing
If config.sh is written with notepad.exe or similar, it will have DOS
newlines. In essence, this adds an extra \r to all variables unless the
user is careful and adds '#' at the end of each line. So, push config.sh
through 'sed' in the container before sourcing the file.

Fixes #2