Remove openbench.exit on launch
A stale /openbench/OpenBench/Client/openbench.exit may be left behind
when the client has been signaled to exit and then killed before the
client exited by itself. This can lead to the client exiting
immediately after the next launch. So, remove this file on launch to
prevent the double exit.