Add support for running the container with --user
To support this, be more precise with the symlinks to /cache, so
that running the OpenBench client never need to write to the container.

The benefit of using --user when launching the container is that it's
now easy to control the uid/gid for accessing the mounted files.

Update also scripts/start-openbench-worker.sh to automatically
set --user to reflect the uid/gid launching the container.